Use the new method in tests
I also renamed the tests to better describe what they're doing
